requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0

# Graph
neo4j>=5.15.0
networkx>=3.2.0
torch-geometric>=2.4.0

# ML
scikit-learn>=1.4.0
xgboost>=2.0.0
torch>=2.2.0

# Data
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0

# Streaming
kafka-python>=2.0.0
redis>=5.0.0

# Entity Resolution
recordlinkage>=0.16
jellyfish>=1.0.0

# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
httpx>=0.27.0
